refatcor: extract header component
