+++ src/app/doctor/page.tsx
@@ -2,7 +2,8 @@
 import React, { useState, useEffect, Suspense } from "react";
 import Image from "next/image";
 import DoctorMap from "@/components/DoctorMap";
-import { useSearchParams, useRouter } from "next/navigation";
+import { useSearchParams } from "next/navigation";
+import Header from "@/components/Header";
 
 interface Doctor {
   id: number;
@@ -63,18 +64,15 @@ const defaultDoctors: Doctor[] = [
 function DoctorPageContent() {
   const searchParams = useSearchParams();
   const [doctors, setDoctors] = useState<Doctor[]>([]);
-  const [searchInput, setSearchInput] = useState("");
-  const [locationInput, setLocationInput] = useState("");
-  const [insuranceInput, setInsuranceInput] = useState("");
-  const [isLoading, setIsLoading] = useState(false);
-  const router = useRouter();
 
   useEffect(() => {
     const doctorsParam = searchParams.get("doctors");
     try {
       if (doctorsParam) {
         const parsed = JSON.parse(decodeURIComponent(doctorsParam));
-        setDoctors(Array.isArray(parsed) ? parsed : parsed.doctors || defaultDoctors);
+        setDoctors(
+          Array.isArray(parsed) ? parsed : parsed.doctors || defaultDoctors
+        );
       } else {
         setDoctors(defaultDoctors);
       }
@@ -83,30 +81,7 @@ function DoctorPageContent() {
     }
   }, [searchParams]);
 
-  const handleSearch = async () => {
-    if (!searchInput.trim()) return;
-    setIsLoading(true);
-    try {
-      const response = await fetch(
-        `https://smarterdoc-backend-1094971678787.us-central1.run.app/v1/search/doctors`,
-        {
-          method: "POST",
-          headers: { "Content-Type": "application/json" },
-          body: JSON.stringify({
-            query: searchInput,
-            location: locationInput,
-            insurance: insuranceInput,
-          }),
-        }
-      );
-      const data = await response.json();
-      router.push(`/doctor?doctors=${encodeURIComponent(JSON.stringify(data.doctors))}`);
-    } catch (err) {
-      console.error(err);
-    } finally {
-      setIsLoading(false);
-    }
-  };
+  
 
   return (
     <main className="relative min-h-screen flex flex-col items-center px-6 py-10">
@@ -119,45 +94,7 @@ function DoctorPageContent() {
         className="object-cover z-0"
       />
 
-      {/* Header */}
-      <header className="flex items-center justify-between w-full max-w-6xl mb-8 z-10">
-        <div className="flex items-center">
-          <Image src="/logo.png" alt="SmartDoc Logo" width={32} height={32} className="mr-2" />
-          <h1 className="text-2xl font-bold text-gray-800">SmarterDoc AI</h1>
-        </div>
-
-        {/* Search bar */}
-        <div className="flex items-center h-12 rounded-full border border-gray-300 bg-white shadow-sm px-6 py-2">
-          <input
-            type="text"
-            placeholder="Search"
-            value={searchInput}
-            onChange={(e) => setSearchInput(e.target.value)}
-            className="flex-1 outline-none bg-transparent text-gray-700 placeholder-gray-400"
-          />
-          <input
-            type="text"
-            placeholder="Location"
-            value={locationInput}
-            onChange={(e) => setLocationInput(e.target.value)}
-            className="flex-1 outline-none bg-transparent text-gray-700 placeholder-gray-400 ml-4 border-l pl-4 border-gray-300"
-          />
-          <input
-            type="text"
-            placeholder="Insurance"
-            value={insuranceInput}
-            onChange={(e) => setInsuranceInput(e.target.value)}
-            className="flex-1 outline-none bg-transparent text-gray-700 placeholder-gray-400 ml-4 border-l pl-4 border-gray-300"
-          />
-          <button
-            onClick={handleSearch}
-            disabled={isLoading}
-            className="flex items-center justify-center h-9 w-9 ml-4 bg-[#433C50] text-white p-2 rounded-full hover:bg-[#5F72BE] transition disabled:opacity-50"
-          >
-            {isLoading ? <i className="ri-loader-4-line animate-spin"></i> : <i className="ri-search-line"></i>}
-          </button>
-        </div>
-      </header>
+      <Header />
 
       {/* Recommended Doctors */}
       <section className="backdrop-blur-md bg-white/60 rounded-3xl shadow-lg p-6 w-full max-w-6xl z-10 mb-10">
