+++ src/app/globals.css
@@ -6,7 +6,7 @@
   --background: #ffffff;
 }
 
-@theme inline {
+:root {
   --color-background: var(--background);
   --color-foreground: var(--foreground);
   --font-sans: var(--font-geist-sans);
